rename graphs file & remove unecesary files
